+++ src/main/deltachat/autocrypt.ts
@@ -1,20 +1,15 @@
 import SplitOut from './splitout'
 import { getLogger } from '../../shared/logger'
+import DeltaChat from 'deltachat-node'
 
 const log = getLogger('main/deltachat/autocrypt')
 
 export default class DCAutocrypt extends SplitOut {
-  initiateKeyTransfer() {
-    return new Promise(resolve => {
-      this._dc.initiateKeyTransfer2((key: string) => {
-        resolve(key)
-      })
-    })
+  initiateKeyTransfer(): Promise<string> {
+    return this._dc.initiateKeyTransfer()
   }
 
-  continueKeyTransfer(messageId: number, setupCode: string) {
-    return new Promise(resolve => {
-      this._dc.continueKeyTransfer2(messageId, setupCode, resolve)
-    })
+  continueKeyTransfer(messageId: number, setupCode: string): Promise<unknown> {
+    return this._dc.continueKeyTransfer(messageId, setupCode)
   }
 }