+++ src/renderer/components/dialogs/Settings-Backup.tsx
@@ -1,11 +1,67 @@
-import React from 'react'
-import { Card, Elevation, H5, Button } from '@blueprintjs/core'
+import React, { useState, useEffect } from 'react'
+import {
+  Card,
+  Elevation,
+  H5,
+  Classes,
+  ProgressBar,
+  Intent,
+  H2,
+} from '@blueprintjs/core'
 import { SettingsButton } from './Settings'
-import { confirmationDialogLegacy as confirmationDialog } from './ConfirmationDialog'
 import { OpenDialogOptions } from 'electron'
 import { ipcBackend } from '../../ipc'
+import { DialogProps } from './DialogController'
+import DeltaDialog, {
+  DeltaDialogBody,
+  DeltaDialogContent,
+  SmallDialog,
+} from './DeltaDialog'
+import { isOpen } from '@blueprintjs/core/lib/esm/components/context-menu/contextMenu'
+import { DeltaProgressBar } from '../Login-Styles'
 
 const { remote } = window.electron_functions
+
+function ExportProgressDialog(props: DialogProps) {
+  const userFeedback = window.__userFeedback
+  const tx = window.translate
+
+  const [progress, setProgress] = useState(0.0)
+
+  const onFileWritten = (_event: any, [_, filename]: [any, string]) => {
+    userFeedback({
+      type: 'success',
+      text: tx('pref_backup_written_to_x', filename),
+    })
+    props.onClose()
+  }
+
+  const onImexProgress = (_: any, [progress, data2]: [number, number]) => {
+    setProgress(progress)
+  }
+  useEffect(() => {
+    ipcBackend.once('DC_EVENT_IMEX_FILE_WRITTEN', onFileWritten)
+    ipcBackend.on('DC_EVENT_IMEX_PROGRESS', onImexProgress)
+
+    return () => {
+      ipcBackend.removeListener('DC_EVENT_IMEX_FILE_WRITTEN', onFileWritten)
+      ipcBackend.removeListener('DC_EVENT_IMEX_PROGRESS', onImexProgress)
+    }
+  }, [])
+
+  return (
+    <SmallDialog isOpen={props.isOpen} onClose={() => {}}>
+      <DeltaDialogBody>
+        <DeltaDialogContent>
+          <H5 style={{ marginTop: '20px' }}>
+            {tx('imex_progress_title_desktop')}
+          </H5>
+          <DeltaProgressBar intent={Intent.PRIMARY} progress={progress} />
+        </DeltaDialogContent>
+      </DeltaDialogBody>
+    </SmallDialog>
+  )
+}
 
 function onBackupExport() {
   const tx = window.translate
@@ -29,16 +85,8 @@ function onBackupExport() {
       }
       remote.dialog.showOpenDialog(opts, (filenames: string[]) => {
         if (!filenames || !filenames.length) return
-        ipcBackend.once('DC_EVENT_IMEX_FILE_WRITTEN', (_event, filename) => {
-          userFeedback({
-            type: 'success',
-            text: tx('pref_backup_written_to_x', filename),
-          })
-
-          closeDialog('ImexProgress')
-        })
+        openDialog(ExportProgressDialog)
         ipcBackend.send('backupExport', filenames[0])
-        openDialog('ImexProgress', {})
       })
     },
   })
@@ -51,7 +99,7 @@ export default function SettingsBackup() {
       <Card elevation={Elevation.ONE}>
         <H5>{tx('pref_backup')}</H5>
         <SettingsButton onClick={onBackupExport}>
-          {tx('pref_backup_export_start_button')}
+          {tx('export_backup_desktop')}
         </SettingsButton>
       </Card>
     </>