+++ CHANGELOG.md
@@ -21,7 +21,8 @@
 - update Readme a bit
 - Fix whitescreen crash on opening the EncryptionInfoDialog
 - Fix drop send multiple files at once ([#1622](https://github.com/deltachat/deltachat-desktop/issues/1622))
-- fix window bounds are not saved correctly #1705
+- Fix window bounds are not saved correctly #1705
+- Fix export backup progress bar missing
 
 ### Changed
 
@@ -37,6 +38,7 @@
 - move the verified icon over the group avatar
 - now it is possible to have multiple accounts with the same email address (you can import a backup next to the active account)
 - change account path again - now the account folder name isn't tied to the email address anymore.
+- Update to deltachat-node v1.35.0 (async core)
 - Add open attachment to message context menu and increase clickable area (click to open attachment)
 
 ### Technical